requirements.txt
bitarray==3.1.0
matplotlib==3.10.1
statsmodels==0.14.4
pandas==2.2.3
mmh3==5.1.0
grpcio==1.60.0
grpcio-tools==1.60.0
msgpack==1.0.7
protobuf==4.25.1